+++ celements-spring-mvc/src/main/java/com/celements/spring/mvc/HelloWorldController.java
@@ -1,40 +1,18 @@
 package com.celements.spring.mvc;
 
-import javax.inject.Inject;
-
-import org.springframework.beans.factory.BeanFactory;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.xwiki.context.Execution;
 
-import com.celements.execution.XWikiExecutionProp;
 import com.celements.spring.security.AuthenticatedBaseController;
-import com.xpn.xwiki.XWikiContext;
 
 @RestController
 public class HelloWorldController extends AuthenticatedBaseController {
 
-  private final BeanFactory beanFactory;
-
-  @Inject
-  public HelloWorldController(BeanFactory beanFactory) {
-    this.beanFactory = beanFactory;
-  }
-
-  @GetMapping("/helloworld")
+  @GetMapping
   @PreAuthorize("permitAll()")
   public String helloWorld() {
     return "Hello World!";
-  }
-
-  @GetMapping("/hellocontext")
-  public XWikiContext helloContext() {
-    return beanFactory
-        .getBean(Execution.class)
-        .getContext()
-        .get(XWikiExecutionProp.XWIKI_CONTEXT)
-        .orElseThrow();
   }
 
 }
